[Refactor] #21 openFeignClient 관련 외부 라이브러리 생성해서 implement (jar 파일)
추후 jitpack 배포 후 교체
