+++ content/build.gradle
@@ -25,6 +25,7 @@ configurations {
 
 repositories {
 	mavenCentral()
+	maven {url 'https://jitpack.io'}
 }
 
 ext {
@@ -58,6 +59,12 @@ dependencies {
 	annotationProcessor "jakarta.persistence:jakarta.persistence-api"
 
 	implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client'
+
+	// myFeignDto jitpack library
+	// implementation 'com.github.bluesun147:myFeignConfig:fb4dd179b7'
+
+	// myFeignDto jar library
+	implementation files('feign-0.0.11.jar')
 }
 
 dependencyManagement {
