+++ content/settings.gradle
@@ -1,1 +1,10 @@
 rootProject.name = 'content-service'
+
+// https://stackoverflow.com/questions/69150086/could-not-find-method-dependencyresolutionmanagement-for-arguments
+//dependencyResolutionManagement {
+//    repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
+//    repositories {
+//        mavenCentral()
+//        maven { url 'https://jitpack.io' }
+//    }
+//}
