+++ content/src/main/java/com/haechan/content/domain/contract/feign/DistributorFeignClient.java
@@ -1,5 +1,6 @@
-package com.haechan.content.global.feign;
+package com.haechan.content.domain.contract.feign;
 
+import com.haechan.feign.dto.DistributorFeignResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
