+++ content/src/main/java/com/haechan/content/domain/contract/service/ContractService.java
@@ -3,12 +3,12 @@ package com.haechan.content.domain.contract.service;
 import com.haechan.content.domain.contract.dto.ContractExcelDto;
 import com.haechan.content.domain.contract.entity.Contract;
 import com.haechan.content.domain.contract.repository.ContractRepository;
+import com.haechan.content.domain.contract.feign.DistributorFeignClient;
 import com.haechan.content.domain.ost.entity.Ost;
 import com.haechan.content.domain.ost.repository.OstRepository;
 import com.haechan.content.global.config.ExcelHelper;
-import com.haechan.content.global.feign.ContractFeignResponse;
-import com.haechan.content.global.feign.DistributorFeignClient;
-import com.haechan.content.global.feign.DistributorFeignResponse;
+import com.haechan.feign.dto.ContractFeignResponse;
+import com.haechan.feign.dto.DistributorFeignResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
