+++ content/src/main/java/com/haechan/content/domain/ost/service/OstService.java
@@ -3,9 +3,11 @@ package com.haechan.content.domain.ost.service;
 import com.haechan.content.domain.ost.dto.OstRegisterDto;
 import com.haechan.content.domain.drama.entity.Drama;
 import com.haechan.content.domain.drama.repository.DramaRepository;
-import com.haechan.content.global.feign.*;
 import com.haechan.content.domain.ost.entity.Ost;
 import com.haechan.content.domain.ost.repository.OstRepository;
+import com.haechan.content.domain.ost.feign.ProducerFeignClient;
+import com.haechan.content.domain.ost.feign.SingerFeignClient;
+import com.haechan.feign.dto.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
