--- content/src/main/java/com/haechan/content/global/feign/ContractFeignResponse.java
@@ -1,14 +1,0 @@
-package com.haechan.content.global.feign;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-@AllArgsConstructor
-public class ContractFeignResponse {
-    private Long contractId;
-    private Double producerPercent;
-    private Double singerPercent;
-}
