+++ content/src/test/java/com/haechan/content/SettlementApplicationTests.java
@@ -1,13 +1,13 @@
-package com.haechan.content;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SettlementApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
+//package com.haechan.content;
+//
+//import org.junit.jupiter.api.Test;
+//import org.springframework.boot.test.context.SpringBootTest;
+//
+//@SpringBootTest
+//class SettlementApplicationTests {
+//
+//	@Test
+//	void contextLoads() {
+//	}
+//
+//}
