+++ content/src/test/java/com/haechan/content/domain/config/TestQueryDslConfig.java
@@ -1,18 +1,18 @@
-package com.haechan.content.domain.config;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-
-@TestConfiguration
-public class TestQueryDslConfig {
-    @PersistenceContext
-    private EntityManager entityManager;
-
-    @Bean
-    public JPAQueryFactory jpaQueryFactory() {
-        return new JPAQueryFactory(entityManager);
-    }
-}
+//package com.haechan.content.domain.config;
+//
+//import com.querydsl.jpa.impl.JPAQueryFactory;
+//import jakarta.persistence.EntityManager;
+//import jakarta.persistence.PersistenceContext;
+//import org.springframework.boot.test.context.TestConfiguration;
+//import org.springframework.context.annotation.Bean;
+//
+//@TestConfiguration
+//public class TestQueryDslConfig {
+//    @PersistenceContext
+//    private EntityManager entityManager;
+//
+//    @Bean
+//    public JPAQueryFactory jpaQueryFactory() {
+//        return new JPAQueryFactory(entityManager);
+//    }
+//}
